refactor: remove global shortcuts, move Shortcuts - update injections in a few components - use Shift+Backspace for back (prevent clashes with the \b) - use string context for single instance components
